simplify code related to ProjectionVsDataBenchmark

+++ src/main/java/com/luxoft/logeek/data/HasIdAndName.java
@@ -1,8 +1,8 @@
 package com.luxoft.logeek.data;
 
 public interface HasIdAndName {
-	
-	Long getId();
-	
-	String getName();
+
+    int getId();
+
+    String getName();
 }
